backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
requests
pymongo
python-dotenv
numpy
